% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacz_kryteria_recznie.R
\name{lacz_kryteria_recznie}
\alias{lacz_kryteria_recznie}
\title{Laczenie kryteriow na podstawie podanych identyfikatorow}
\usage{
lacz_kryteria_recznie(skale, laczenia, src = NULL)
}
\arguments{
\item{skale}{wektor liczbowy z id_skali lub ciąg znaków z wyrażeniem
regularnym identyfikującymi skale po kolumnie 'opis'}

\item{laczenia}{lista wektorów tekstowych - każdy element list zawiera nazwy
kryteriów (w formie 'k_idKryterium'), które powinny zostać połączone w jedno
pseudokryterium
\code{\link[EWDdane]{pobierz_wyniki_surowe}}}

\item{src}{NULL połączenie z bazą danych IBE zwracane przez funkcję
\code{\link[ZPD]{polacz}}. Jeśli nie podane, podjęta zostanie próba
automatycznego nawiązania połączenia.}
}
\value{
data frame
Kolumna \code{elementy} zawiera data frame'y które mogą zostać użyte
jako argument funkcji \code{\link[ZPDzapis]{edytuj_skale}}.
}
\description{
Funkcja łączy, w ramach podanych skal, podane kryteria oceny poszczególnych
zadań w pseudokryteria. Jest ona użyteczna w szczególności do skal
opisujących wyniki egzaminu na wejściu do skalowania na potrzeby modeli
latentnej EWD, gdzie określenie kryteriów do połączenia odbywa się na
podstawie innych skal (obejmujących tylko dany egzamin w danym roku), trzeba
jednak potem móc te łączenia zastosować do skal łączących wyniki egzaminu
z dwóch kolejnych lat (dla danego typu szkoły).
}
