% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacz_kryteria_z_korelacji.R
\name{lacz_kryteria_z_korelacji}
\alias{lacz_kryteria_z_korelacji}
\title{Laczenie kryteriow na podstawie korelacji polichorycznych}
\usage{
lacz_kryteria_z_korelacji(
  skale,
  katalogDane = "dane surowe/",
  prog = 0.5,
  tylkoWWiazkach = TRUE
)
}
\arguments{
\item{skale}{wektor liczbowy z id_skali lub ciąg znaków z wyrażeniem
regularnym identyfikującymi skale po kolumnie 'opis'}

\item{katalogDane}{ciąg znaków - ścieżka do katalogu, w którym znajdują
się dane z wynikami surowymi egzaminów, pobranymi przy pomocy funkcji
\code{\link[EWDdane]{pobierz_wyniki_surowe}}}

\item{prog}{liczbowa z zakresu (0;1) - minimalna wartość korelacji, do której
ma być kontynuowane łączenie}

\item{tylkoWWiazkach}{wartość logiczna - czy dopuszczać łączenie kryteriów
tylko w ramach wiązek pytań/kryteriów?}
}
\value{
lista, której każdy element jest dwuelemntową listą, zawierającą id
skali oraz data frame, której można użyć jako argument \code{elementy}
funkcji \code{\link[ZPDzapis]{edytuj_skale}}
}
\description{
Funkcja łączy, w ramach podanych skal, kryteria oceny poszczególnych zadań
w pseudokryteria. Łączenie odbywa się na podstawie wartości korelacji
polichorycznych, obliczanych przy pomocy funkcji
\code{\link[polycor]{polychor}} z pakietu \code{polycor}.
}
