% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lacz_kryteria_z_korelacji.R
\name{lacz_kryteria_z_korelacji_w_ramach_skali}
\alias{lacz_kryteria_z_korelacji_w_ramach_skali}
\title{Laczenie kryteriow na podstawie korelachi polichorycznych}
\usage{
lacz_kryteria_z_korelacji_w_ramach_skali(x, katalogDane, prog, tylkoWWiazkach)
}
\arguments{
\item{x}{data frame opisujący kryteria oceny w ramach skali}

\item{katalogDane}{ciąg znaków - ścieżka do katalogu, w którym znajdują
się dane z wynikami surowymi egzaminów, pobranymi przy pomocy funkcji
\code{\link[EWDdane]{pobierz_wyniki_surowe}}}

\item{prog}{liczbowa z zakresu (0;1) - minimalna wartość korelacji, do której
ma być kontynuowane łączenie}

\item{tylkoWWiazkach}{wartość logiczna - czy dopuszczać łączenie kryteriów
tylko w ramach wiązek pytań/kryteriów?}
}
\value{
data frame, której można uzyć jako argument \code{elementy}
funkcji \code{\link[ZPDzapis]{edytuj_skale}}
}
\description{
Koń roboczy wywoływany przez \code{\link{lacz_kryteria_z_korelacji}}.
}

