% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacz_kryteria_z_nr_zadan.R
\name{lacz_kryteria_z_nr_zadan_w_ramach_skali}
\alias{lacz_kryteria_z_nr_zadan_w_ramach_skali}
\title{Laczenie kryteriow na podstawie numerow zadan}
\usage{
lacz_kryteria_z_nr_zadan_w_ramach_skali(x)
}
\arguments{
\item{x}{data frame opisujący kryteria oceny w ramach skali}
}
\value{
data frame, której można uzyć jako argument \code{elementy}
funkcji \code{\link[ZPDzapis]{edytuj_skale}}
}
\description{
Koń roboczy wywoływany przez \code{\link{lacz_kryteria_z_nr_zadan}}.
}
