% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pobierz_kryteria_do_laczenia.R
\name{pobierz_kryteria_do_laczenia}
\alias{pobierz_kryteria_do_laczenia}
\title{Laczenie kryteriow}
\usage{
pobierz_kryteria_do_laczenia(skale)
}
\arguments{
\item{skale}{wektor liczbowy z id_skali lub ciąg znaków z wyrażeniem
regularnym identyfikującymi skale po kolumnie 'opis'}
}
\value{
data table
}
\description{
Na podstawie informacji pozwalających zidentyfikować skale funkcja
przygotowuje obiekt z kryteriami oceny, który obrabiają potem funkcje
\code{\link{lacz_kryteria_z_nr_zadan}} lub
\code{\link{lacz_kryteria_z_korelacji}}.
}

