% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_polaczone_kryteria.R
\name{pobierz_polaczone_kryteria}
\alias{pobierz_polaczone_kryteria}
\title{Skracanie skal oceny}
\usage{
pobierz_polaczone_kryteria(skale, echo = FALSE, src = NULL)
}
\arguments{
\item{skale}{wektor liczbowy z id_skali lub ciąg znaków z wyrażeniem
regularnym identyfikującymi skale po kolumnie 'opis'}

\item{echo}{wartość logiczna - czy wypisać na konsoli kod pozwalający
zdefiniować zwracaną listę?}

\item{src}{NULL połączenie z bazą danych IBE zwracane przez funkcję
\code{\link[ZPD]{polacz}}. Jeśli nie podane, podjęta zostanie próba
automatycznego nawiązania połączenia.}
}
\value{
Lista wektorów tekstowych - każdy element zawiera nazwy kryteriów
oceny tworzących jedno pseudokryterium - które mogą zostać użyte
jako argument \code{laczenia} funkcji \code{\link{lacz_kryteria_recznie}}.
}
\description{
Funkcja pobiera definicje pseudokryteriów przypisanych do podanych skal
i zwraca je w formie, w jakiej mogą zostać przekazane do
\code{\link{lacz_kryteria_recznie}}.
}
