% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przewidywanie_rasch.R
\name{przewidywanie_rasch}
\alias{przewidywanie_rasch}
\title{Oszacowania z modeli Rascha}
\usage{
przewidywanie_rasch(sumy, oszacowania, maks = NULL, span = 0.2)
}
\arguments{
\item{sumy}{data frame zawierający id_obserwacji i obliczone sumy punktów
(może to być kilka kolumn z sumami z kilku różnych części egzaminu)}

\item{oszacowania}{data frame zawierający id_obserwacji i oszacowania
umiejętności z modelu}

\item{maks}{opcjonalnie wektor liczb zawierających maksymalne możliwe do
uzyskania wartości sum - jego elementy muszą nazywać się tak, jak kolumny
argumentu \code{sumy}, opisujące wartości sum punktów}

\item{span}{wartość liczbowa - parametr \code{span}, przekazywany do funkcji
\code{loess()} przy wygładzaniu i zapełnianiu dziur w przebiegu przekodowania}
}
\value{
lista o elementach:
\itemize{
  \item{\code{mapowanie} Data frame zawierający mapowanie sum na średnie
        oszacowania. Kolumny o nazwach takich, jak kolumny w argumencie
        \code{sumy} (nie przechowujące id obserwacji) przyjmują wartość
        \code{TRUE} jeśli dany wiersz dotyczy posiadających wynik
        z odpowiedniej części egzaminu lub wartość \code{FALSE}, jeśli dany
        wiersz dotyczy zdających, którzy nie posiadają wyniku z odpowiedniej
        części egzaminu. Kolumna \code{suma} opisuje sumę punktów łącznie
        z wszystkich części (z których ktoś posiada wyniki), a ostatnia
        kolumna zawiera wartość przewidywania.}
  \item{\code{przewidywania} Data frame zawierający zastosowanie ww.
        mapowania do danych, z którymi wywołana została funkcja.}
  \item{\code{odsUtraconejWariancji} Odsetek wariancji oszacowań cechy
        utracony w wyniku uśredniania.}
}
}
\description{
Funkcja przygotowuje przeliczenie z sumy punktów na przewidywanie poziomu
umiejętności z modelu IRT (żeby miało to większy sens powinna to być jakaś
odmiana modelu Rascha).
}
\details{
W modelu Rascha liczba punktów w zasadzie powinna być statystyką dostateczną
dla oszacowania poziomu umiejętności bez względu na to, czy test składa się
wyłącznie z zadań ocenianych binarnie, czy też zawiera również zadania
o większej liczbie poziomów wykonania. Niestety, w Mplusie tak nie jest
i jeśli test zawiera zadania o większe liczbie poziomów wykonania, Mplus
będzie przypisywał zdającym nieco inne oszacowania w zależności od tego,
które konkretnie zadania rozwiązali poprawnie, a które niepoprawnie.

Jeśli napodstawie wyników działania Mplusa chcemy uzyskać funkcję
przypisującą jednoznacznie oszacowanie sumie punktów, musimy w związku z tym
dokonać jakiegoś przybliżenia. Możliwe są dwa podejścia:
\itemize{
  \item{Wygenerować wszystkie możliwe wektory odpowiedzi, na podstawie
        wyestymowanych parametrów modelu obliczyć odpowiadające im
        oszacowania oraz ich wiarygodność, a następnie agregować w ramach
        grup wyróżnionych ze względu na sumę punktów, z użyciem średniej
        ważonej wiarygodnością profilu (a więc wartością proporcjonalną do
        teoretycznej częśtości jego występowania).}
  \item{Użyć oszacowań obliczonych na bardzo dużym zbiorze danych
        (rzeczywistym lub wygenerowanym metodą Monte-Carlo) i obliczyć
        średnie (już nieważone) oszacowań w ramach grup o tej samej wartości
        sumy.}
}
Podejście pierwsze jest teoretycznie bardziej poprawne, ale znajduje
zastosowanie tylko dla dosyć krótkich testów, jako że liczba wszystkich
możliwych wektorów odpowiedzi rośnie wykładniczo wraz ze wzrostem liczby
zadań. Podejście drugie jest aplikowalne zawsze, ale jego sensowność silnie
zależy od iczby losowanych wyników (procedura Monte-Carlo) lub liczby
obserwacji i rozkładu wyników (wykorzystanie danych rzeczywistych).

Niniejsza funkcja implementuje drugą procedurą, w oparciu o dane rzeczywiste
(czyli w zasadzie najbardziej ryzykowną). Stara się przy tym wykryć
i naprawić pewne problemy, które mogą się z nią wiązać:
\itemize{
  \item{niemonotoniczność przewidywań,}
  \item{dziury w rozkładzie sumy.}
}
Rozwiązanie obu problemów stara się uzyskać wygładzając i interpoulując
przewidywanie przy pomocy regresji nieparametrycznej (funkcją \code{loess()},
wywoływana z parametrem \code{span}). Jeśli przewidywanie wynikające
z regesji nieparametrycznej jest niemonotoniczne, funkcja będzie zwiększać
wartość parametru span o 0.05 tak długo, aż przewidywanie stanie się
monotoniczne (wtedy zwórcone zostanie ostrzeżenie) lub aż osiągnieta zostanie
wartość parametru \code{span} większa lub równa 1 (jeśli przewidywanie cały
czas będzie niemonotonicze, funkcja zwróci błąd).

Jeśli minimalna wartość występująca w danych nie jest równa 0, to wartościom
niższym od najmniejszej występującej (w ramach danej grupy) przypisane
zostanie przewidywanie odpowiadające tej minimalnej wartości sumy,
występującej w danych (w ramach danej grupy).

Analogicznie, jeśli podana została wartość parametru \code{maks}, to w ramach
każdej grupy ew. wartościom większym od maksymalnej wartości występującej
w danych (w ramach danej grupy) a nie większym od odpowiedniej wartości
parametru \code{maks} przypisane zostaną przewidywania odpowiadające tej
maksymalnej wartości sumy, występującej w danych (w ramach danej grupy).
}
