% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rysuj_laczenia_z_korelacji.R
\name{rysuj_laczenia_z_korelacji_w_ramach_czesci}
\alias{rysuj_laczenia_z_korelacji_w_ramach_czesci}
\title{Laczenie kryteriow na podstawie korelacji polichorycznych}
\usage{
rysuj_laczenia_z_korelacji_w_ramach_czesci(x, wielkoscTekstu = 12)
}
\arguments{
\item{x}{pojedynczy wiersz obiektu klasy \code{wynikLaczKryteriaZKorelacji}}

\item{wielkoscTekstu}{bazowa wielkość tekstu w \code{pts}}
}
\value{
data frame
}
\description{
Koń roboczy dla \code{\link{rysuj_laczenia_z_korelacji}}.
}
