% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skaluj_matura.R
\name{skaluj_matura}
\alias{skaluj_matura}
\title{Procedury skalowania egzaminow}
\usage{
skaluj_matura(rok, processors = 2, opis = "skalowanie do EWD",
  katalogSurowe = "../../dane surowe",
  katalogWyskalowane = "../../dane wyskalowane", zapisz = TRUE,
  skala = NULL, proba = -1, usunWieleNaraz = FALSE)
}
\arguments{
\item{rok}{rok przeprowadzenie egzaminu}

\item{processors}{liczba rdzeni do wykorzystania przy estymacji}

\item{opis}{opcjonalnie ciąg znaków - opis skalowania}

\item{katalogSurowe}{opcjonalnie ścieżka do katalogu, w którym znajdują się
pliki z zapisanymi (przy pomocy funkcji
\code{\link[EWDdane]{pobierz_wyniki_surowe}} z pakietu EWDdane) surowymi
wynikami egzaminu}

\item{katalogWyskalowane}{opcjonalnie ścieżka do katalogu, w którym znajdują
się pliki z zapisanymi (przy pomocy funkcji
\code{\link[EWDdane]{pobierz_wyniki_wyskalowane}} z pakietu EWDdane)
wyskalowanymi wynikami egzaminu}

\item{zapisz}{wartość logiczna - czy zapisać wyniki do pliku .RData?}

\item{skala}{id_skali (liczba naturalna) lub ciąg znaków z wyrażeniem
regularnym, do którego ma pasować opis skali}

\item{proba}{opcjonalnie liczba natrualna - wielkość próby, jaka ma być
wylosowana z danych przed estymacją modelu; przydatne (tylko) do testów
działania funkcji}

\item{usunWieleNaraz}{opcjonalnie wartość logiczna - jeśli wiele
(pseudo)kryteriów oceny ma dyskryminację poniżej 0,2, to czy usuwać je
wszystkie w jednym kroku?}
}
\value{
lista klasy \code{listaWynikowSkalowania}, której elementy są listami
klasy \code{wynikiSkalowania} i składają się z elementów:
\itemize{
  \item{\code{skalowania} data frame o kolumnach:
        \itemize{
          \item{\code{skalowanie,}}
          \item{\code{opis,}}
          \item{\code{estymacja,}}
          \item{\code{id_skali,}}
          \item{\code{do_prezentacji,}}
          \item{\code{data;}}
        }}
  \item{\code{skalowania_grupy} data frame o kolumnach:
        \itemize{
          \item{\code{id_skali,}}
          \item{\code{skalowanie,}}
          \item{\code{grupa;}}
        }}
  \item{\code{skalowania_elementy} data frame o kolumnach:
        \itemize{
          \item{\code{id_skali,}}
          \item{\code{kolejnosc,}}
          \item{\code{skalowanie,}}
          \item{\code{parametr,}}
          \item{\code{model,}}
          \item{\code{wartosc,}}
          \item{\code{uwagi,}}
          \item{\code{bs,}}
          \item{\code{grupowy,}}
          \item{\code{grupa;}}
        }}
  \item{\code{skalowania_obserwacje} data frame o kolumnach:
        \itemize{
          \item{\code{id_skali,}}
          \item{\code{skalowanie,}}
          \item{\code{id_obserwacji,}}
          \item{\code{id_testu,}}
          \item{\code{estymacja,}}
          \item{\code{nr_pv,}}
          \item{\code{wynik,}}
          \item{\code{bs,}}
          \item{\code{grupa;}}
        }}
  \item{\code{skalowania} data frame o kolumnach:
        \itemize{
          \item{\code{skalowanie,}}
          \item{\code{opis,}}
          \item{\code{estymacja,}}
          \item{\code{id_skali,}}
          \item{\code{do_prezentacji,}}
          \item{\code{data;}}
        }}
  \item{\code{usunieteKryteria} wektor tekstowy z nazwami (pseudo)kryteriów,
        które zostały usunięte podczas skalowania wzorcowego;}
}
}
\description{
Funkcja przeprowadza skalowanie czterech wskaźników osiągnięć maturalnych:
1) humanistycznego (j. polski, historia, WOS), 2) matematyczno-przyrodniczego
(matematyka, biologia, chemia, fizyka, geografia, informatyka),
3) polonistycznego (j. polski), 4) matematycznego (matematyka), na potrzeby
wyliczania trzyletnich wskaźników dla LO i techników. Wykorzystywane są
wielogrupowe modele 2PL/SGRM. Grupy definiowane są albo przez wybór poziomu
rozszerzonego i typ szkoły (LO/T) - dla wskaźników 1), i 4), albo przez
typ szkoły (LO/T) - dla wskaźników 2), i 3).
}
\details{
\bold{Uwaga}, oszacowania zwracane przez funkcję \bold{nie są porównywalne
pomiędzy LO a T!}
}
\seealso{
\code{\link[EWDskalowanie]{skaluj}},
\code{\link[EWDskalowanie]{procedura_1k_1w}},
\code{\link{sprawdz_wyniki_skalowania}}
}

