% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skaluj_matura_rasch.R
\name{skaluj_matura_rasch}
\alias{skaluj_matura_rasch}
\title{Procedury skalowania egzaminow}
\usage{
skaluj_matura_rasch(rok, processors = 2,
  opis = "skalowanie do Kalkulatora EWD",
  katalogSurowe = "../../dane surowe",
  katalogWyskalowane = "../../dane wyskalowane", zapisz = TRUE,
  skala = NULL, proba = -1)
}
\arguments{
\item{rok}{rok przeprowadzenie egzaminu}

\item{processors}{liczba rdzeni do wykorzystania przy estymacji}

\item{opis}{opcjonalnie ciąg znaków - opis skalowania}

\item{katalogSurowe}{opcjonalnie ścieżka do katalogu, w którym znajdują się
pliki z zapisanymi (przy pomocy funkcji
\code{\link[EWDdane]{pobierz_wyniki_surowe}} z pakietu EWDdane) surowymi
wynikami egzaminu}

\item{katalogWyskalowane}{opcjonalnie ścieżka do katalogu, w którym znajdują
się pliki z zapisanymi (przy pomocy funkcji
\code{\link[EWDdane]{pobierz_wyniki_wyskalowane}} z pakietu EWDdane)
wyskalowanymi wynikami egzaminu}

\item{zapisz}{wartość logiczna - czy zapisać wyniki do pliku .RData?}

\item{skala}{id_skali (liczba naturalna) lub ciąg znaków z wyrażeniem
regularnym, do którego ma pasować opis skali}

\item{proba}{opcjonalnie liczba natrualna - wielkość próby, jaka ma być
wylosowana z danych przed estymacją modelu; przydatne (tylko) do testów
działania funkcji}
}
\value{
lista klasy \code{listaWynikowSkalowania}, której elementy są listami
klasy \code{wynikiSkalowania} i składają się z elementów:
\itemize{
  \item{\code{skalowania} data frame o kolumnach:
        \itemize{
          \item{\code{skalowanie,}}
          \item{\code{opis,}}
          \item{\code{estymacja,}}
          \item{\code{id_skali,}}
          \item{\code{do_prezentacji,}}
          \item{\code{data;}}
        }}
  \item{\code{skalowania_grupy} data frame o kolumnach:
        \itemize{
          \item{\code{id_skali,}}
          \item{\code{skalowanie,}}
          \item{\code{grupa;}}
        }}
  \item{\code{skalowania_elementy} data frame o kolumnach:
        \itemize{
          \item{\code{id_skali,}}
          \item{\code{kolejnosc,}}
          \item{\code{skalowanie,}}
          \item{\code{parametr,}}
          \item{\code{model,}}
          \item{\code{wartosc,}}
          \item{\code{uwagi,}}
          \item{\code{bs,}}
          \item{\code{grupowy,}}
          \item{\code{grupa;}}
        }}
  \item{\code{skalowania_obserwacje} data frame o kolumnach:
        \itemize{
          \item{\code{id_skali,}}
          \item{\code{skalowanie,}}
          \item{\code{id_obserwacji,}}
          \item{\code{id_testu,}}
          \item{\code{estymacja,}}
          \item{\code{nr_pv,}}
          \item{\code{wynik,}}
          \item{\code{bs,}}
          \item{\code{grupa;}}
        }}
  \item{\code{skalowania} data frame o kolumnach:
        \itemize{
          \item{\code{skalowanie,}}
          \item{\code{opis,}}
          \item{\code{estymacja,}}
          \item{\code{id_skali,}}
          \item{\code{do_prezentacji,}}
          \item{\code{data;}}
        }}
  \item{\code{normy} data frame o kolumnach:
        \itemize{
          \item{\code{id_skali,}}
          \item{\code{skalowanie,}}
          \item{\code{grupa,}}
          \item{\code{wartosc,}}
          \item{\code{wartosc_zr;}}
       }}
  \item{\code{usunieteKryteria} wektor tekstowy z nazwami (pseudo)kryteriów,
        które zostały usunięte podczas skalowania wzorcowego;}
}
}
\description{
Funkcja przeprowadza skalowanie skalowanie matury z matematyki łącznie na
poziomach podstawowym i rozszerzonym, z wykorzystaniem wielogrupowego modelu
Rascha (na potrzeby maturalnego Kalkulatora EWD). Grupy definiowane są przez
wybór poziomu rozszerzonego i typ szkoły (LO/T).
}
\details{
Schemat przekodowania sum punktów na oszacowania umiejętności obliczany jest
na podstawie danych wzorcowych, przy pomocy funkcji
\code{\link{przewidywanie_rasch}}, a następnie na jego podstawie przypisywane
są wartości przewidywane wszystkim zdającym.

\bold{Uwaga}, oszacowania i normy zwracane przez funkcję \bold{nie są
porównywalne pomiędzy LO a T!}
}
\seealso{
\code{\link[EWDskalowanie]{skaluj}},
\code{\link[EWDskalowanie]{procedura_1k_1w}},
\code{\link{sprawdz_wyniki_skalowania}}
}

