% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skroc_skale_oceny.R
\name{skroc_skale_oceny}
\alias{skroc_skale_oceny}
\title{Skracanie skal oceny}
\usage{
skroc_skale_oceny(
  skale,
  katalogDane = "dane surowe/",
  maxLPozWyk = 5,
  minLiczebnPozWyk = 100,
  minOdsPozWyk = 0.05,
  print = TRUE,
  populacja = c("wy", "we"),
  src = NULL
)
}
\arguments{
\item{skale}{wektor liczbowy z id_skali lub ciąg znaków z wyrażeniem
regularnym identyfikującymi skale po kolumnie 'opis'}

\item{katalogDane}{ciąg znaków - ścieżka do katalogu, w którym znajdują
się dane z wynikami surowymi egzaminów, pobranymi przy pomocy funkcji
\code{\link[EWDdane]{pobierz_wyniki_surowe}}}

\item{maxLPozWyk}{maksymalna liczba poziomów pytania}

\item{minLiczebnPozWyk}{minimalna liczebność obserwacji w każdym z poziomów}

\item{minOdsPozWyk}{minimalny odsetek obserwacji, które ma zawierać każdy
poziom}

\item{print}{wartość logiczna - czy pokazywać informacje o skracaniu?}

\item{populacja}{ciąg znaków określający, czy przy obliczaniu rozkładów
mają zostać uwzględnione obserwacje spełniające kryterium przynależności
do \emph{populacji na wyjściu} (domyślnie) czy do \emph{populacji na wejściu}}

\item{src}{NULL połączenie z bazą danych IBE zwracane przez funkcję
\code{\link[ZPD]{polacz}}. Jeśli nie podane, podjęta zostanie próba
automatycznego nawiązania połączenia.}
}
\value{
data frame
Kolumna \code{elementy} zawiera data frame'y które mogą zostać użyte
jako argument funkcji \code{\link[ZPDzapis]{edytuj_skale}}.
}
\description{
Funkcja dokonuje skrótu skali oceny (pseudo)kryteriów w oparciu o rozkład
ich wyników w grupie kalibracyjnej i informację o wszystkich możliwych
wartościach, jakie może przyjąć dane (pseudo)kryterium.
}
