% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/skracanie_poziomy.R
\name{skroc_skale_oceny_w_ramach_skali}
\alias{skroc_skale_oceny_w_ramach_skali}
\title{Skracanie skal oceny}
\usage{
skroc_skale_oceny_w_ramach_skali(x, katalogDane = "../dane surowe/",
  maxLPozWyk = 5, minLiczebnPozWyk = 100, minOdsPozWyk = 0.05,
  print = TRUE)
}
\arguments{
\item{x}{pojedynczy wiersz data frame'a, opisujący skalę
regularnym identyfikującymi skale po kolumnie 'opis'}

\item{katalogDane}{ciąg znaków - ścieżka do katalogu, w którym znajdują
się dane z wynikami surowymi egzaminów, pobranymi przy pomocy funkcji
\code{\link[EWDdane]{pobierz_wyniki_surowe}}}

\item{maxLPozWyk}{maksymalna liczba poziomów pytania}

\item{minLiczebnPozWyk}{minimalna liczebność obserwacji w każdym z poziomów}

\item{minOdsPozWyk}{minimalny odsetek obserwacji, które ma zawierać każdy
poziom}

\item{print}{wartość logiczna - czy pokazywać informacje o skracaniu?}
}
\value{
Data frame, pasujący swoją strukturą jako argument \code{elementy}
do funkcji \code{edytuj_skale} z pakietu \code{ZPD} lub lista takich data
frame'ów.
}
\description{
Koń roboczy wywoływany przez \code{\link{skroc_skale_oceny}}.
}

