% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprawdz_wyniki_skalowania.R
\name{sprawdz_wyniki_skalowania}
\alias{sprawdz_wyniki_skalowania}
\title{Diagnostyka skalowania}
\usage{
sprawdz_wyniki_skalowania(nazwaPliku)
}
\arguments{
\item{nazwaPliku}{ciąg znaków - nazwa pliku .RData, zawierającego wyniki
skalowania (typowo zapisanego przez funkcję \code{\link{skaluj_spr}},
\code{\link{skaluj_egz_gimn}}, \code{\link{skaluj_egz_gimn_rasch}},
\code{\link{skaluj_matura}} lub \code{\link{skaluj_matura_rasch}}).}
}
\value{
funkcja nie zwraca żadnych wartości
}
\description{
Funkcja wyrysowuje zestaw wykresów diagnostycznych pozwalających przyjrzeć
się wynikom uzyskanym z modelu skalowania.
}
