% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytaj_parametry_UIRT.R
\name{wczytaj_parametry_UIRT}
\alias{wczytaj_parametry_UIRT}
\title{Wczytuje parametry skalowania przeprowadzonego UIRT-em}
\usage{
wczytaj_parametry_UIRT(
  sciezkaWe,
  src = NULL,
  sciezkaWy = NULL,
  skalowanie = NULL,
  opis = NULL
)
}
\arguments{
\item{sciezkaWe}{ścieżka do pliku CSV z parametrami skalowania
przeprowadzonego UIRT-em}

\item{src}{NULL połączenie z bazą danych IBE zwracane przez funkcję
\code{\link[ZPD]{polacz}}. Jeśli nie podane, podjęta zostanie próba
automatycznego nawiązania połączenia.}

\item{sciezkaWy}{NULL ścieżka do pliku .RData, w którym zapisane zostaną
wyniki. Jeśli NULL, wynik nie zostanie zapisany do pliku.}

\item{skalowanie}{NULL jeśli podane, nadpisuje id skalowania wczytane z pliku
wejściowego (uwaga, ma sens raczej tylko gdy plik wejściowy opisuje tylko
jedno skalowanie)}

\item{opis}{NULL jeśli podany, nadpisuje opis skalowania wczytany z pliku
wejściowego (uwaga, ma sens raczej tylko gdy plik wejściowy opisuje tylko
jedno skalowanie)}
}
\value{
lista opisująca wyniki skalowania w formacie zgodnym z tym, jakiego
  używa unkcja \code{\link[ZPDzapis]{zapisz_skalowanie}}.
}
\description{
Wczytuje parametry z pliku CSV i konwertuje je na strukturę
  danych zgodną z tą, jakiej używa funkcja
  \code{\link[ZPDzapis]{zapisz_skalowanie}}.

  Opcjonalnie zapisuje wynik konwersji do pliku .RData.
}
