% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytaj_dane.R
\name{wczytaj_wyniki_surowe}
\alias{wczytaj_wyniki_surowe}
\title{Wczytywanie wynikow egzaminow zapisanych na dysku}
\usage{
wczytaj_wyniki_surowe(katalogDane, rodzajEgzaminu, czescEgzaminu, rok,
  idSkali, kryteria = NULL)
}
\arguments{
\item{katalogDane}{ciąg znaków - ścieżka do katalogu, w którym znajdują
się dane z wynikami surowymi egzaminów, pobranymi przy pomocy funkcji
\code{\link[EWDdane]{pobierz_wyniki_surowe}}}

\item{rodzajEgzaminu}{ciąg znaków}

\item{czescEgzaminu}{ciąg znaków}

\item{rok}{liczba naturalna}

\item{idSkali}{liczba naturalna - id_skali w bazie dla skali, która ma zostać
zastosowana do wyników surowych}

\item{kryteria}{opcjonalnie wektor tekstowy z nazwami wszystkich
kryteriów oceny, jakie powinna zawierać dana część/ci egzaminu}
}
\value{
data frame (data table)
}
\description{
Funkcja wczytuje wyniki surowe egzaminu, zapisane wcześniej na dysku
funkcją \code{\link[EWDdane]{pobierz_wyniki_surowe}} z pakietu EWDdane.
}
\details{
Funkcja co do zasady dołącza do wyników egzaminu dane kontekstowe, zawężając
grupę zwracanych obserwacji do tych, dla których te dane istnieją. Jeśli
jednak stwierdzi, że w efekcie usunięci zostaliby wszyscy, to zwróci wyniki,
bez dołączania do nich danych kontekstowych. Dojdzie do tego w szczególności
przy wczytywaniu wyników egzaminu gimnazjalnego z lat wcześniejszych niż
2006 r., gdyż są to dane CKE i dane pobierane funkcją
\code{\link[EWDdane]{pobierz_dane_kontekstowe}} ich nie obejmują.
}
