% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wczytaj_dane.R
\name{wczytaj_wyniki_wyskalowane}
\alias{wczytaj_wyniki_wyskalowane}
\title{Wczytywanie wynikow egzaminow zapisanych na dysku}
\usage{
wczytaj_wyniki_wyskalowane(katalogDane, rodzajEgzaminu, rok, idSkali)
}
\arguments{
\item{katalogDane}{ciąg znaków - ścieżka do katalogu, w którym znajdują
się dane z wyskalowanymi wynikami egzaminów, pobranymi przy pomocy funkcji
\code{\link[EWDdane]{pobierz_wyniki_wyskalowane}}}

\item{rodzajEgzaminu}{ciąg znaków}

\item{rok}{liczba naturalna}

\item{idSkali}{liczba naturalna}
}
\value{
data frame (data table)
}
\description{
Funkcja wczytuje wyskalowane wyniki egzaminu, zapisane wcześniej na dysku
funkcją \code{\link[EWDdane]{pobierz_wyniki_wyskalowane}} z pakietu EWDdane.
}

