% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wylicz_normy_ekwikwantylowe.R
\name{wylicz_normy_ekwikwantylowe}
\alias{wylicz_normy_ekwikwantylowe}
\title{Obliczanie norm ekwikwantylowych}
\usage{
wylicz_normy_ekwikwantylowe(egzamin, rok, katalogZDanymi, ...)
}
\arguments{
\item{egzamin}{ciąg znaków: "sprawdzian" | "egzamin gimnazjalny" | "matura"}

\item{rok}{liczba całkowita - rok, dla którego mają zostać obliczone normy}

\item{katalogZDanymi}{ciąg znaków - ścieżka do katalogu, w którym znajdują
się pliki .RData z wynikami surowymi egzaminów i danymi kontekstowymi
o uczniach i szkołach ściągnięte wcześniej przy pomocy funkcji
\code{\link[EWDdane]{pobierz_wyniki_surowe}}}

\item{...}{ew. parametry przekazywane do funkcji
\code{\link[ZPD]{normy_ekwikwantylowe}}}
}
\value{
data table
}
\description{
Funkcja służy do obliczenia norm ekwikwantylowych na podstawie wyników
egzaminów pobranych wcześniej przy pomocy funkcji
\code{\link[EWDdane]{pobierz_wyniki_surowe}}.
}
\details{
Nominalnie działa dla wszystkich egzaminów, ale ponieważ w obecnej wersji nie
udostępnia grupowania, więc w odniesieniu do matury jest mało użyteczna
(niemniej obecnie normalizacja ekwikwantylowa nie jest wykorzystywana
w odniesieniu do wyników matury).

Dodatkowe uproszczenie polega na tym, że w zwracanym data frame'ie kolumna
\code{skalowanie} jest zawsze wypełniona wartością \code{1}, co w ogólności
nie zawsze musi być słuszne.
}

