% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacz_kryteria_recznie.R
\name{zdefiniuj_laczenia}
\alias{zdefiniuj_laczenia}
\title{Laczenie kryteriow na podstawie podanych identyfikatorow}
\usage{
zdefiniuj_laczenia(kryteriaSkali, laczenia)
}
\arguments{
\item{kryteriaSkali}{data frame opisujący kryteria oceny w ramach
pojedynczej skali}

\item{laczenia}{lista wektorów tekstowych - każdy element list zawiera nazwy
kryteriów (w formie 'k_idKryterium'), które powinny zostać połączone w jedno
pseudokryterium}
}
\value{
data frame
}
\description{
Koń roboczy wywoływany przez
\code{\link{lacz_kryteria_recznie}}.
}
