% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rysuj_laczenia_z_korelacji.R
\name{rysuj_laczenia_z_korelacji}
\alias{rysuj_laczenia_z_korelacji}
\title{Laczenie kryteriow na podstawie korelacji polichorycznych}
\usage{
rysuj_laczenia_z_korelacji(x, wielkoscTekstu = 1, prog = NULL)
}
\arguments{
\item{x}{obiekt klasy \code{wynikLaczKryteriaZKorelacji}}

\item{wielkoscTekstu}{bazowa wielkość tekstu w \code{pts}}

\item{prog}{opcjonalnie liczba z przedziału (0;1), wskazująca, że narysowane
mają być łączenia od pierwszego do tego, które poprzedza pierwsze łączenie
wykonane przy wartości korelacji poniżej zadanego progu}
}
\value{
data frame
Kolumna \code{elementy} zawiera listy data frame'ów, które mogą zostać użyte
jako argument funkcji \code{\link[ZPDzapis]{edytuj_skale}}. Pierwszy element
opisuje sytuację bez łączenia (a więc de facto brak zmian w skali -
w zakresie danej części egzaminu), a kolejne odpowiednio kolejne łączenia.
}
\description{
Funkcja wizualizuje skutki łączeń kryteriów oceny, będących efektem użycia
funkcji \code{\link{lacz_kryteria_z_korelacji}} dla dyskryminacji zadań.
}
