% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rysuj_laczenia_z_korelacji.R
\name{rysuj_laczenia_z_korelacji}
\alias{rysuj_laczenia_z_korelacji}
\title{Laczenie kryteriow na podstawie korelacji polichorycznych}
\usage{
rysuj_laczenia_z_korelacji(x)
}
\arguments{
\item{x}{obiekt klasy \code{wynikLaczKryteriaZKorelacji}}
}
\value{
data frame
}
\description{
Funkcja wizualizuje skutki łączeń kryteriów oceny, będących efektem użycia
funkcji \code{\link{lacz_kryteria_z_korelacji}} dla dyskryminacji zadań.
}

