% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rysuj_laczenia_z_korelacji.R
\name{rysuj_laczenia_z_korelacji_w_ramach_skali}
\alias{rysuj_laczenia_z_korelacji_w_ramach_skali}
\title{Laczenie kryteriow na podstawie korelacji polichorycznych}
\usage{
rysuj_laczenia_z_korelacji_w_ramach_skali(x)
}
\arguments{
\item{x}{pojedynczy wiersz obiektu klasy \code{wynikLaczKryteriaZKorelacji}}
}
\value{
data frame
}
\description{
Koń roboczy dla \code{\link{rysuj_laczenia_z_korelacji}}.
}

