% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/skaluj_egz_gimn.R
\name{skaluj_egz_gimn}
\alias{skaluj_egz_gimn}
\title{Procedury skalowania egzaminow}
\usage{
skaluj_egz_gimn(rok, processors = 2, katalogSurowe = "../../dane surowe",
  katalogWyskalowane = "../../dane wyskalowane", zapisz = TRUE,
  skala = NULL)
}
\arguments{
\item{rok}{rok przeprowadzenie egzaminu}

\item{processors}{liczba rdzeni do wykorzystania przy estymacji}

\item{katalogSurowe}{opcjonalnie ścieżka do katalogu, w którym znajdują się
pliki z zapisanymi (przy pomocy funkcji
\code{\link[EWDdane]{pobierz_wyniki_surowe}} z pakietu EWDdane) surowymi
wynikami egzaminu}

\item{katalogWyskalowane}{opcjonalnie ścieżka do katalogu, w którym znajdują
się pliki z zapisanymi (przy pomocy funkcji
\code{\link[EWDdane]{pobierz_wyniki_wyskalowane}} z pakietu EWDdane)
wyskalowanymi wynikami egzaminu}

\item{zapisz}{wartość logiczna - czy zapisać wyniki do pliku .RData?}

\item{skala}{id_skali (liczba naturalna) lub ciąg znaków z wyrażeniem
regularnym, do którego ma pasować opis skali}
}
\value{
lista klasy \code{listaWynikowSkalowania}, której elementy są listami
klasy \code{wynikiSkalowania} i składają się z elementów:
\itemize{
  \item{\code{idSkali} id_skali w bazie;}
  \item{\code{skalowanie} nr skalowania w bazie;}
  \item{\code{usunieteKryteria} wektor tekstowy z nazwami (pseudo)kryteriów, które
        zostały usunięte podczas skalowania wzorcowego;}
  \item{\code{parametry} data frame z wyestymowanymi parametrami modelu w jego
        ostatecznej postaci (tj. takiej, jak w ostatnim kroku skalowania wzorcowego
        i w jedynym kroku skalowania na wszystkich zdających);}
  \item{\code{oszacowania} data frame zawierający id_obserwacji i wyliczone
        oszacowania umiejętności dla wszystkich zdających;}
  \item{\code{rzetelnoscEmpiryczna} rzetelność wyliczona na podstawie oszacowań ze
        skalowania wzorcowego (jako wariancja oszacowań EAP);}
}
}
\description{
Funkcja przeprowadza skalowanie wyników egzaminu gimnazjalnego.
}
\seealso{
\code{\link[EWDskalowanie]{skaluj}},
\code{\link[EWDskalowanie]{procedura_1k_1w}},
\code{\link{sprawdz_wyniki_skalowania}}
}

