% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sprawdz_wyniki_skalowania.R
\name{sprawdz_wyniki_skalowania_konstruktu}
\alias{sprawdz_wyniki_skalowania_konstruktu}
\title{Diagnostyka skalowania}
\usage{
sprawdz_wyniki_skalowania_konstruktu(model)
}
\arguments{
\item{model}{obiekt klasy \code{wynikiSkalowania}}
}
\value{
funkcja nie zwraca żadnych wartości
}
\description{
Funkcja wyrysowuje zestaw wykresów diagnostycznych pozwalających przyjrzeć
się wynikom uzyskanym z modelu skalowania.
}

