% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprawdz_wyniki_skalowania.R
\name{sprawdz_wyniki_skalowania_konstruktu}
\alias{sprawdz_wyniki_skalowania_konstruktu}
\title{Diagnostyka skalowania}
\usage{
sprawdz_wyniki_skalowania_konstruktu(model, src = NULL)
}
\arguments{
\item{model}{obiekt klasy \code{wynikiSkalowania}}

\item{src}{NULL połączenie z bazą danych IBE zwracane przez funkcję
\code{\link[ZPD]{polacz}}. Jeśli nie podane, podjęta zostanie próba
automatycznego nawiązania połączenia.}
}
\value{
funkcja nie zwraca żadnych wartości
}
\description{
Funkcja wyrysowuje zestaw wykresów diagnostycznych pozwalających przyjrzeć
się wynikom uzyskanym z modelu skalowania.
}
