% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmien_parametry_na_do_bazy.R
\name{zmien_parametry_na_do_bazy}
\alias{zmien_parametry_na_do_bazy}
\title{Konwersja tabeli parametrow na format bazy}
\usage{
zmien_parametry_na_do_bazy(x, idSkali, skalowanie,
  rzetelnoscEmpiryczna = NULL, standaryzacja = NULL, grupy = NULL)
}
\arguments{
\item{x}{data frame z wartościami parametrów w formacie zwracanym przez
funkcję \code{\link[EWDskalowanie]{skaluj}}}

\item{idSkali}{liczba naturalna - id_skali w bazie}

\item{skalowanie}{liczba naturalna - nr skalowania w bazie}

\item{rzetelnoscEmpiryczna}{opcjonalnie liczba lub data frame - rzetelność
empiryczna testu (w przypadku data frame'a - w poszczególnych grupach)}

\item{standaryzacja}{opcjonalnie data frame o kolumnach \code{grupa},
\code{sr}, \code{os}, zawierający parametry standaryzacji oszacowań
w poszczególnych grupach}

\item{grupy}{data frame zawierający mapowanie numerów grup na ich nazwy);
musi składać się z dwóch zmiennych: 'grupa', zawierającej nazwy grup
i drugiej, o dowolnej nazwie, zawierającej numery grup}
}
\value{
ramka danych w formacie odpowiadającym strukturze tablicy
\code{skalowania_elementy} w bazie
}
\description{
Funkcja przekształca tabelę wartości parametrów modelu skalowania z formatu
zwracanego przez funkcję \code{\link[EWDskalowanie]{skaluj}} na format,
w jakim wartości parametrów zapisywane są w bazie i na klasyczną,
jednowymiarowąz parametryzację IRT.
}
